Add Newtonsoft.Json
Since Newtonsoft.Json is already a .NET Core based project, we can use
the dotnet tooling that we use to run our outer build to do the build
of Newtonsoft.Json itself.

I've introduced a new property that will be the path to the .NET Core
tooling (so we can swap it out when we eventually do the two phase
build using the .NET Core CLI that we just built).