FromSource => ArcadeBuildFromSource
Switch to a more qualified prop name that clearly ties it to Arcade.

FromSource is already used here in the dotnet org:
https://github.com/dotnet/maui/blob/c7e2c55ace3ec55a0858f1c4bb55476375ea72f8/Environment.Build.props#L29
and it's the name of a .NET API. This doesn't force us to change, but it
shows it's not necessarily a great/searchable term.

Also add more info to the readme.