Fix build against musl libc (#318)
Originally reported at https://github.com/dotnet/installer/issues/12859

Use the same logic as dotnet-isntall.sh to identify whether a system is
musl-based or not. For musl-based Linux systems, use `linux-musl` as the
platform name, which leads to better RID names.

Otherwise, we end up trying to use artifacts with an incorrect RID on
musl based systems.

In particular, we look for `runtime.linux-x64.microsoft.netcore.ildasm`
which doesn't work on musl-based systems. It's also not produced by
source-build builds, which makes building using a previous build of
source-build impossible.