Add pakrym/jab as a module and repo project
Jab today is the only solution to use AOT source generated dependency
injection which avoids Reflection and better supports NativeAOT. As the
package is a build only dependency, it won't be listed as a package
dependency anywhere but is a required build dependency.

Eventually in the future, when dotnet/runtime offers a solution for
source generated dependency injection, this repository can be removed.

Jab will be used in the Compatibility tooling in dotnet/sdk.