Add xunit 2.4.2 and xunit-abstractions 2.0.3 (#167)
* Add xunit 2.4.2 and xunit-abstractions 2.0.3
This PR adds xunit and xunit abstractions into source-build-externals. The set of functionality source-built for xunit is relatively limited:
- TFMs are limited to netstandard2.0
- Packages include:
  - xunit.core (text only package)
  - xunit.abstractions
  - xunit.extensibility.core
  - xunit.extensibility.execution
  - xunit.runner.utility

The changes required for the projects generally centered around:
- Trimming/updating of TFMs
- Removal of automated versioning logic
- Trimming away of the projects to build

In addition to that, I had to make some small tweaks to the way the repo builds. Because xunit depends on xunit.abstractions, to avoid prebuilts we need to introduce a DependencyPackageProjects style 2-phase build.

* Check out submodules recursively