General SBRP infrastructure improvements (#1186)
* Disable NuGetAudit and NuGet package pruning warnings in GenAPI wrapper project

* Use TaskHostFactory in PackageSourceGenerator

Without TaskFactory="TaskHostFactory", the live built of PackageSourceGeneratorTask won't be used in a build loaded from a running node.

* Use required modified in msbuild task

* Remove handling of framework assemblies

The framework assemblies concept in a nuspec isn't used by .NET Standard or .NETCoreApp packages. It's mainly needed for .NET Framework which this repo doesn't have assemblies for.

* Remove NETStandard.Library implicit package reference

That was needed for netstandard1.x packages which this repo doesn't support anymore.