Normalize casing of targeting pack projects and nuspecs
All SBRP projects except the targeting packs consistently use the casing
of the package id (+ "." + packageversion), i.e.
`Microsoft.Extensions.Logging.8.0.0.csproj`.

A few targeting pack projects didn't follow that casing. Similar for the
nuspecs which should always be lower-cased.

This is in preparation to use the live targeting packs in SBRP
and remove the self SBRP dependency.